Fixed password verification
During registration, we confirm passwords. For some reason, using the logic we currently had, even if the password was the exact same as the confirm password, the error that pops up when you don't have the same password pops up. This is because of the logic used for password verification where apparently if you use concatenation like previously instead of .equals for string comparison, there are weird complications. Fixed so now registration works normally.