Added some extra functions
Added some functionality like role selection in areas (such as, when you login you can specify how to log in as, admin, student, or instructor which is checks in databaseHelper login. Login is still a bit wonky, for some reason the boolean for login being successful doesnt work. Also added back button to Register to take you back to login.