Added dropTables for second table (articleList) and cleaned up code
If you have to add/remove a column on an SQL database, you have to first remove the original database and then create it again. I added a drop table option for the second database so it can be updated. make sure to comment them both out if you want the database to remain after the application closes.